refactor: race service를 준비, 진행, 결과로 나누기
