+++ src/main/java/racingcar/race/RaceController.java
@@ -8,13 +8,17 @@ import java.util.Scanner;
 
 public class RaceController {
     private ParticipatingCars cars;
-    private RaceService raceService;
+    private RacePreparation racePreparation;
+    private RaceProcess raceProcess;
+    private RaceResult raceResult;
     private int raceRound;
     private int leadPosition = 0;
 
     public RaceController() {
         this.cars = new ParticipatingCars();
-        this.raceService = new RaceService();
+        this.racePreparation = new RacePreparation();
+        this.raceProcess = new RaceProcess();
+        this.raceResult = new RaceResult();
     }
 
     public void generateCars(Scanner scanner) {
@@ -22,8 +26,8 @@ public class RaceController {
             System.out.println(Constant.PARTICIPATING_CAR_NAME_INPUT_MESSAGE);
 
             try {
-                String[] allCarName = raceService.getAllCarName(scanner);
-                raceService.carRegistration(cars, allCarName);
+                String[] allCarName = racePreparation.getAllCarName(scanner);
+                racePreparation.carRegistration(cars, allCarName);
                 break;
             } catch (IllegalArgumentException e) {
                 System.out.println(e.getMessage());
@@ -36,7 +40,7 @@ public class RaceController {
             System.out.println(Constant.RACE_NUMBER_INPUT_MESSAGE);
 
             try {
-                raceRound = raceService.getRaceRound(scanner);
+                raceRound = racePreparation.getRaceRound(scanner);
                 break;
             } catch (IllegalArgumentException e) {
                 System.out.println(e.getMessage());
@@ -48,13 +52,13 @@ public class RaceController {
         System.out.println(Constant.RACE_PROCESS);
 
         for (int i = 0; i < raceRound; i++) {
-            this.leadPosition = raceService.driveCar(cars, leadPosition);
-            raceService.showCurrentSituation(cars);
+            this.leadPosition = raceProcess.driveCar(cars, leadPosition);
+            raceProcess.showCurrentSituation(cars);
         }
     }
 
     public void showWinner() {
-        List<String> winners = raceService.findWinner(cars, leadPosition);
-        raceService.printWinner(winners);
+        List<String> winners = raceResult.findWinner(cars, leadPosition);
+        raceResult.printWinner(winners);
     }
 }
